requirements.txt
# Dependências principais
streamlit==1.31.1
fastapi==0.110.1
uvicorn==0.34.0

# Supabase e conexões
supabase==2.15.0
psycopg2==2.9.10
SQLAlchemy<2.0

# Utilitários
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.2.0
tiktoken==0.5.2

# Weaviate compatível com seu código
weaviate-client==4.10.0

# Protobuf compatível com gRPC e streamlit
protobuf>=5.26.1,<6.0


langchain-community==0.0.33
